Restructured configuration to clarify target vs context
A target represents a deployed UAA. A context is the user/client/target
combination used to obtain a token with a particular set of permissions.  An
operator using the cli might have multiple UAAs that he wants to switch between
with multiple contexts for each target.

Building for this use case isn't a super high priority, but I felt like I was
referring to the context and config objects so much that I knew I wanted to do
this restructuring now before the code gets too large to feasibly change.