Rip out the client autoapprove field
This field has so many confusing issues associated with it that I don't want to
spend further effort figuring out how to parse and display and re-serialize
these values to the user. A field that can be bool or []string or an array of
stringified boolean values provides no useful information to an operator.