Allow overriding the uaa-cli config directory with an environment variable
By default, uaa-cli config file will be stored in `$HOME/.uaa/config.json`, but
if the environment variable `UAA_HOME` is set, then the config file will be
stored there instead.

This allows for some flexibility in how users of the tool use it, in the same
ways that cf-cli does.